requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
datasets
